+++ FirebaseAI/CHANGELOG.md
@@ -1,6 +1,13 @@
 # 12.2.0
 - [feature] Added support for returning thought summaries, which are synthesized
   versions of a model's internal reasoning process. (#15096)
+- [feature] Added a new configuration option to use limited-use App
+  Check tokens for attesting Firebase AI Logic requests. This enhances
+  security against replay attacks. To use this feature, configure it
+  explicitly via the new `useLimitedUseAppCheckTokens` parameter when
+  initializing `FirebaseAI`. We recommend migrating to limited-use
+  tokens now, so your app will be ready to take advantage of replay
+  protection when it becomes available for Firebase AI Logic. (#15099)
 
 # 12.0.0
 - [feature] Added support for Grounding with Google Search. (#15014)